Added hook useDynamicStyles and separated Typography component
