+++ frontend/src/components/common/Typography/types.ts
@@ -1,0 +1,20 @@
+import { CSSProperties, ReactNode } from 'react';
+
+export type TypographyProps = {
+  children: ReactNode;
+  id?: string;
+  className?: string;
+  size?: 'x-small' | 'small' | 'medium' | 'large' | 'x-large';
+  align?: 'left' | 'right' | 'center';
+  color?: 'dark' | 'white' | 'blue';
+  bold?: boolean;
+  lineHeight?: string | number;
+  textTransform?: 'none' | 'capitalize' | 'uppercase' | 'lowercase';
+  textDecoration?: 'none' | 'underline' | 'overline' | 'line-through';
+  letterSpacing?: string | number;
+  display?: 'block' | 'inline' | 'inline-block';
+  margin?: string | number;
+  padding?: string | number;
+  ellipsis?: boolean;
+  additionalStyles?: CSSProperties;
+};
